requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
numpy==2.2.6
oauthlib==3.2.2
opencv-python==4.11.0.86
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
pyzbar==0.1.9
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
soupsieve==2.7
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
